% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpumf_calc.R
\name{cpumf_total}
\alias{cpumf_total}
\alias{cpumf_mean}
\title{Total likelihood with replicate weights}
\usage{
cpumf_total(r)

cpumf_mean(x, w, r)
}
\description{
Total likelihood with replicate weights

Calculate mean likelihood with replicate weights
}
