% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censuspumf.R, R/zzz.R
\docType{data}
\name{records}
\alias{records}
\alias{var_labels}
\alias{encode_labels}
\title{Census PUMF data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 980868 rows and 129 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 144 rows and 2 columns.
}
\usage{
records

var_labels

encode_labels(d)
}
\description{
Census PUMF data

Variable labels
}
\keyword{datasets}
